+++ packages/local_auth/local_auth_platform_interface/lib/local_auth_platform_interface.dart
@@ -89,6 +89,12 @@ abstract class LocalAuthPlatform extends PlatformInterface {
     throw UnimplementedError('isDeviceSupported() has not been implemented.');
   }
 
+  /// Returns false if user change biometric value you must to authorize the new biometric.
+  Future<bool> isValidBiometricAuthorized() async {
+    throw UnimplementedError(
+        'isValidBiometricAuthorized() has not been implemented.');
+  }
+
   /// Cancels any authentication currently in progress.
   ///
   /// Returns true if auth was cancelled successfully.