+++ packages/local_auth/local_auth_windows/lib/local_auth_windows.dart
@@ -65,4 +65,8 @@ class LocalAuthWindows extends LocalAuthPlatform {
   /// Always returns false as this method is not supported on Windows.
   @override
   Future<bool> stopAuthentication() async => false;
+
+  @override
+  Future<bool> isValidBiometricAuthorized() async =>
+      _api.isValidBiometricAuthorized();
 }