+++ packages/local_auth/local_auth_windows/lib/src/messages.g.dart
@@ -47,6 +47,33 @@ class LocalAuthApi {
     }
   }
 
+  /// Returns false if user change biometric value you must to authorize the new biometric.
+  Future<bool> isValidBiometricAuthorized() async {
+    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
+        'dev.flutter.pigeon.LocalAuthApi.isValidBiometricAuthorized', codec,
+        binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel.send(null) as List<Object?>?;
+    if (replyList == null) {
+      throw PlatformException(
+        code: 'channel-error',
+        message: 'Unable to establish connection on channel.',
+      );
+    } else if (replyList.length > 1) {
+      throw PlatformException(
+        code: replyList[0]! as String,
+        message: replyList[1] as String?,
+        details: replyList[2],
+      );
+    } else if (replyList[0] == null) {
+      throw PlatformException(
+        code: 'null-error',
+        message: 'Host platform returned null value for non-null return value.',
+      );
+    } else {
+      return (replyList[0] as bool?)!;
+    }
+  }
+
   /// Attempts to authenticate the user with the provided [localizedReason] as
   /// the user-facing explanation for the authorization request.
   ///