+++ packages/local_auth/local_auth_android/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.33
+- Added PackageManager in native Android code, to correctly prioritize and identify available biometrics modes.
+- Added isValidBiometricAuthorized, should be used when you want to authorize register device biometric.
+- Change authorize method to invalidate when user change biometric device.
+
 ## 1.0.32
 
 * Fixes stale ignore: prefer_const_constructors.