+++ packages/local_auth/local_auth_ios/CHANGELOG.md
@@ -3,6 +3,11 @@
 * Fixes stale ignore: prefer_const_constructors.
 * Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
 
+## 1.1.4
+
+- Added isValidBiometricAuthorized, should be used when you want to authorize register device biometric.
+- Change authorize method to invalidate when user change biometric device.
+
 ## 1.1.3
 
 * Migrates internal implementation to Pigeon.